# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Apployd Control Plane API
  version: 0.1.0
  description: REST API for account, projects, deployments, billing, and observability.
servers:
  - url: http://localhost:4000/api/v1
paths:
...
  /projects:
    get:
      tags: [Projects]
      security: [{ bearerAuth: [] }]
      summary: List projects by organization
      parameters:
        - in: query
          name: organizationId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Project list }
    post:
      tags: [Projects]
      security: [{ bearerAuth: [] }]
      summary: Create project with resource allocation
      responses:
        '201': { description: Project created }
  /projects/{projectId}/resources:
    patch:
      tags: [Projects]
      security: [{ bearerAuth: [] }]
      summary: Update project resource allocation
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Updated }
  /projects/{projectId}/secrets:
    get:
      tags: [Secrets]
      security: [{ bearerAuth: [] }]
      summary: List secret keys for project
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Secret metadata }
  /projects/{projectId}/secrets/{key}:
    put:
      tags: [Secrets]
      security: [{ bearerAuth: [] }]
      summary: Upsert encrypted secret
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
        - in: path
          name: key
          required: true
          schema: { type: string }
      responses:
        '201': { description: Secret stored }
    delete:
      tags: [Secrets]
      security: [{ bearerAuth: [] }]
      summary: Delete secret
      responses:
        '200': { description: Secret deleted }
...
  /usage/summary:
    get:
      tags: [Usage]
      security: [{ bearerAuth: [] }]
      summary: Show pooled usage for current billing period
      parameters:
        - in: query
          name: organizationId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Usage snapshot }
  /usage/daily:
    get:
      tags: [Usage]
      security: [{ bearerAuth: [] }]
      summary: Daily aggregate usage for one metric
      parameters:
        - in: query
          name: organizationId
          required: true
          schema: { type: string }
        - in: query
          name: metricType
          required: true
          schema:
            type: string
            enum: [cpu_millicore_seconds, ram_mb_seconds, bandwidth_bytes, request_count]
        - in: query
          name: days
          required: false
          schema: { type: integer, minimum: 1, maximum: 90 }
      responses:
        '200': { description: Daily usage series }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT